requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
opencv-python>=4.6.0
numpy>=1.21.0